HBX-1919: Move 'org.hibernate.tool.internal.reveng.BinderUtils' to 'org.hibernate.tool.internal.reveng.binder.BinderUtils'
